% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aa.colors.R
\docType{data}
\name{aa.col}
\alias{aa.col}
\title{Color Look-up-Table}
\format{\preformatted{'data.frame':	36 obs. of  4 variables:
 $ nt    : chr  "red" "blue" "orchid" "orchid" ...
 $ aa    : chr  "#FFCC33" "#CC3333" "#3333CC" "#3333CC" ...
 $ charge: chr  "grey42" "grey42" "red" "red" ...
 $ taylor: chr  "#CCFF00" "#FFFF00" "#FF0000" "#FF0066" ...
}}
\usage{
aa.col
}
\description{
Used internally to depict amino acid mutations.
To customize the color scheme, edit 'color-lut.csv',
then uncomment the 'overwrite' argument, rebuild, reinstall, and restart.
}
\keyword{datasets}

