% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.timepoint.parser.R
\name{create.timepoint.parser}
\alias{create.timepoint.parser}
\title{Generate a closure function to parse timepoint labels from sequence names.}
\usage{
create.timepoint.parser(field.sep, field.num)
}
\arguments{
\item{field.sep}{Character used to separate fields in sequence names.}

\item{field.num}{Field in sequence names that contains timepoint labels.}
}
\value{
Validated function to return the specific, delimited field from each sequence name.
}
\description{
Generate a closure function to parse timepoint labels from sequence names.
}
\examples{
tp <- create.timepoint.parser(field.sep=".", field.num=1)
tp <- create.timepoint.parser(field.sep="_", field.num=2)
}

