% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/env.features.R
\docType{data}
\name{env.features}
\alias{env.features}
\title{Env Feature Annotation}
\format{\preformatted{ logi [1:856, 1:21] FALSE FALSE FALSE FALSE FALSE FALSE ...
 - attr(*, "dimnames")=List of 2
  ..$ : chr [1:856] "1" "2" "3" "4" ...
  ..$ : chr [1:21] "CD4" "PGT121" "PGT121" "PGT127" ...
}}
\source{
\url{http://hiv.lanl.gov/content/sequence/HIV/MAP/Env_features.xls}
}
\usage{
env.features
}
\description{
Identifies sites in HIV-1 Env associated with neutralizating antibody responses,
whether by structural contact or site-directed mutagenesis with phenotypic assay.
}
\keyword{datasets}

