% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.fasta.swarmset.R
\name{export.fasta.swarmset}
\alias{export.fasta.swarmset}
\title{Output fasta-formatted full-length sequence text for all selected sequences in swarmset.}
\usage{
export.fasta.swarmset(x, f)
}
\arguments{
\item{x}{swarmset}

\item{f}{Output file name.}
}
\description{
Output fasta-formatted full-length sequence text for all selected sequences in swarmset.
}
\seealso{
\code{\link{swarmset}}

Other swarmset methods: \code{\link{export.swarmset}},
  \code{\link{make.timepoint.logos}},
  \code{\link{plot.swarmset}},
  \code{\link{print.swarmset}},
  \code{\link{resample.swarmset}},
  \code{\link{summary.swarmset}}, \code{\link{swarmset}},
  \code{\link{tabulate.swarmset}}
}
