% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.swarmset.R
\name{export.swarmset}
\alias{export.swarmset}
\title{Output full-length sequence text for all selected sequences in swarmset.}
\usage{
export.swarmset(x)
}
\arguments{
\item{x}{swarmset}

\item{...}{not currently used but could and should be soon}
}
\description{
Output full-length sequence text for all selected sequences in swarmset.
}
\seealso{
\code{\link{swarmset}}

Other swarmset methods: \code{\link{export.fasta.swarmset}},
  \code{\link{make.timepoint.logos}},
  \code{\link{plot.swarmset}},
  \code{\link{print.swarmset}},
  \code{\link{resample.swarmset}},
  \code{\link{summary.swarmset}}, \code{\link{swarmset}},
  \code{\link{tabulate.swarmset}}
}
