% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_help.R
\name{get_help}
\alias{get_help}
\title{Help Understanding Errors and Warnings}
\usage{
get_help(...)
}
\arguments{
\item{\ldots}{Code that is giving an error or warning.}
}
\description{
This function walks a user through the process of understanding errors and
warnings; providing applicable StackOverflow and GitHub or Google searches.
Last, helps the user to format a proper StackOverflow question or GitHub
issues post.
}
\note{
This function cannot help with errors stemming from R's parser
rejecting the code as valid code. The error will occur before our function
will have been called. If such an error occurs, it means the code is not
well-formed. The following SO link gives common errors that fall into the
parse error class: \url{http://stackoverflow.com/q/25889234/1000343}.
}
\examples{
\dontrun{
p <- 1:10
get_help(mean(LETTERS))

get_help(3*"d")

a1 <- c("a","a","b","b","c","d","e","e")
b2 <- c("01.01.2015", "02.02.2015", "14.02.2012", "16.08.2008",
    "17.06.2003", "31.01.2015", "07.01.2022", "09.05.2001")
c3 <- c("1a", "2b", "3c", "4d", "5e", "6f", "7g", "8h")
d3 <- c(1:8)

df2 <- data.frame(a1,b2,c3,d3, stringsAsFactors = F)

library(dplyr)
library(magrittr)

get_help(
  df2 \%>\%
    group_by(a1) \%>\%
    as.Date(b2, format = "\%d.\%m.\%Y")
)
}
}
\keyword{help}
\keyword{question}

