% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.protein.R
\name{is.protein}
\alias{is.protein}
\title{verify that the input alignment consists of amino acids, using relative frequency of iupac nts
this logic is adapted from Will Fischer's pixel routine... but were did he get it?}
\usage{
is.protein(alignment_matrix)
}
\description{
verify that the input alignment consists of amino acids, using relative frequency of iupac nts
this logic is adapted from Will Fischer's pixel routine... but were did he get it?
}
\keyword{internal}
