% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lassie-package.R
\docType{package}
\name{lassie}
\alias{lassie}
\alias{lassie-package}
\alias{package-lassie}
\title{What the Package Does (one line)}
\description{
What the package does (one or more paragraphs)
}

