% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.swarmset.R
\name{plot.swarmset}
\alias{plot.swarmset}
\title{Render a logo plot of concatamer forms of clones in working_swarm}
\usage{
\method{plot}{swarmset}(SST, format = "png", aspect_ratio = 3)
}
\arguments{
\item{SST}{A swarmset object populated with working_swarm.}

\item{format}{Image format.}

\item{aspect_ratio}{Aspect ratio (width to height) of image.}
}
\value{
An explicit path to the file generated, located in a directory removed at the end of the R session.
}
\description{
Render a logo plot of concatamer forms of clones in working_swarm
}
\seealso{
Other swarmset.methods: \code{\link{print.swarmset}};
  \code{\link{resample.swarmset}};
  \code{\link{summary.swarmset}}; \code{\link{swarmset}}
}

