% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.swarmtools.R
\name{plot.swarmtools}
\alias{plot.swarmtools}
\title{Plot a cumulative distribution function of number of sites selected vs. TF loss cutoff.}
\usage{
\method{plot}{swarmtools}(S)
}
\arguments{
\item{S}{Swarmtools object}
}
\description{
Plot a cumulative distribution function of number of sites selected vs. TF loss cutoff.
}

