% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pngs2o.R
\name{pngs2o}
\alias{pngs2o}
\title{Given amino acid sequence/s in one of several formats, replaces N 
(asparagine) with an O character to depict potential N-linked glycosylation 
(PNG) motifs.}
\usage{
pngs2o(x = NULL)
}
\arguments{
\item{x}{an input character string or matrix of aligned amino-acid characters.}
}
\value{
Output type should match input type, with Ns in PNG motifs renamed to O.
}
\description{
NB: Assumes without testing that the input is amino acids, not DNA.
}
\details{
The PNG motif is NxS or NxT, where x is any amino acid except
proline.  Skips gap characters (taken to be the dash or hyphen, i.e. '-').
}
\examples{
lassie::pngs2o("NNST-N-PN-S-T")
lassie::pngs2o(letters)
lassie::pngs2o("no png sites here")

}

