% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.aln.R
\name{prep.aln}
\alias{prep.aln}
\title{Prepare input data for analysis after defining refseq, running read.alignment.file and set.tf if needed}
\usage{
prep.aln(S)
}
\description{
Prepare input data for analysis after defining refseq, running read.alignment.file and set.tf if needed
}
\keyword{internal}
