% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.swarmset.R
\name{print.swarmset}
\alias{print.swarmset}
\title{List concatamers of selected sites for all working_swarm sequences in swarmset.}
\usage{
\method{print}{swarmset}(x, ...)
}
\arguments{
\item{x}{swarmset}

\item{...}{not currently used by could and should be soon}
}
\description{
List concatamers of selected sites for all working_swarm sequences in swarmset.
}
\seealso{
\code{\link{swarmset}}

Other swarmset.methods: \code{\link{make.timepoint.logos}};
  \code{\link{plot.swarmset}};
  \code{\link{resample.swarmset}};
  \code{\link{summary.swarmset}}; \code{\link{swarmset}}
}

