% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.swarmtools.R
\name{print.swarmtools}
\alias{print.swarmtools}
\title{Print a table of the selected sites.}
\usage{
\method{print}{swarmtools}(S)
}
\arguments{
\item{Swarmtools}{object}
}
\description{
Print a table of the selected sites.
}

