% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.alignment.file.R
\name{read.alignment.file}
\alias{read.alignment.file}
\title{Read a fasta-formatted alignment file}
\usage{
read.alignment.file(aas_file = NULL, aas_prefix = NULL,
  alignment_format = "fasta")
}
\arguments{
\item{aas_file}{File containing aligned protein sequences}

\item{aas_prefix}{Optional alternative to specify input alignment (see examples)}

\item{alignment_format}{Format of alignment file/s; must be one of these: \code{fasta}, \code{clustal}, \code{phylip}, \code{msf}, or \code{mase}.}
}
\value{
A matrix that contains the alignment.
}
\description{
Read a fasta-formatted alignment file
}
\examples{
\dontrun{
Equivalent filenames, because the default format is fasta.
aln1 <- read.alignment.file("foo.fasta")
aln2 <- read.alignment.file(aas_prefix="foo")
}
}

