% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/resample.swarmset.R
\name{resample.swarmset}
\alias{resample.swarmset}
\title{Randomly sample as many sequences as are in the given working swarmset, using the same inclusion/exclusion constraints, for a fair comparison.}
\usage{
resample.swarmset(SST, n_replicates = 100)
}
\arguments{
\item{SST}{swarmset object}

\item{n_replicates}{number of replicates}
}
\description{
Randomly sample as many sequences as are in the given working swarmset, using the same inclusion/exclusion constraints, for a fair comparison.
}
\seealso{
\code{\link{swarmset.default}}

Other swarmset.methods: \code{\link{plot.swarmset}};
  \code{\link{print.swarmset}};
  \code{\link{summary.swarmset}};
  \code{\link{swarmset.default}}
}

