% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.alignment.R
\name{set.alignment}
\alias{set.alignment}
\title{Defines alignment file but does not read it.  After setting it, call prep.aln.}
\usage{
set.alignment(S, a, f)
}
\arguments{
\item{S}{swarmtools object}

\item{aln_file}{alignment file}

\item{alignment_format}{Alignment format}
}
\description{
Reading the alignment file is done by prep.aln for control of excise.refseq status.
}
\keyword{internal}

