% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.tf.loss.cutoff.R
\name{set.tf.loss.cutoff}
\alias{set.tf.loss.cutoff}
\title{Define the threshold value, above which sites are considered selected}
\usage{
set.tf.loss.cutoff(S, c)
}
\arguments{
\item{S}{swarmtools object}

\item{c}{cutoff threshold value, from 0 to 100}
}
\value{
swarmtools object with list of sites in selected_sites if possible
}
\description{
Define the threshold value, above which sites are considered selected
}

