% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set.tf.R
\name{set.tf}
\alias{set.tf}
\title{Accessor method to update tf-derived variables (tf_loss, is_tf, peak_tf_loss) when set.}
\usage{
set.tf(S)
}
\description{
Accessor method to update tf-derived variables (tf_loss, is_tf, peak_tf_loss) when set.
}
\keyword{internal}

