% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.tp.parser.R
\name{set.tp.parser}
\alias{set.tp.parser}
\title{Update parser function, which extracts timepoint labels from sequence names.}
\usage{
set.tp.parser(S, f)
}
\arguments{
\item{S}{swarmtools object}

\item{f}{timepoint parser function, created by create.timepoint.parser().}
}
\value{
swarmtools object with updated parser.  If an alignment file is defined, it will be reparsed.
}
\description{
Update parser function, which extracts timepoint labels from sequence names.
}
\seealso{
\code{\link{create.timepoint.parser}}
}

