% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.swarmset.R
\name{summary.swarmset}
\alias{summary.swarmset}
\title{List names of sequences in working_swarm of swarmset.}
\usage{
\method{summary}{swarmset}(S)
}
\arguments{
\item{S}{swarmset}
}
\description{
List names of sequences in working_swarm of swarmset.
}
\seealso{
\code{\link{swarmset.default}}

Other swarmset.methods: \code{\link{plot.swarmset}};
  \code{\link{print.swarmset}};
  \code{\link{resample.swarmset}};
  \code{\link{swarmset.default}}
}

