% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.swarmtools.R
\name{summary.swarmtools}
\alias{summary.swarmtools}
\title{Summarize selected sites.}
\usage{
\method{summary}{swarmtools}(object, ...)
}
\arguments{
\item{object}{Swarmtools object}

\item{...}{not currently in use but should be soon}
}
\value{
A data frame of selected sites, with one row per ?
}
\description{
Summarize selected sites.
}
