% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarmtools.R
\name{swarmtools}
\alias{swarmtools}
\title{Identify sites under putative immune selection.}
\usage{
swarmtools(aas_aln = NULL, aas_file = NULL,
  alignment_format = "fasta", tf_index = 1, tf_name = NULL,
  timepoints_parser = NULL, refseq_lut_file = NULL,
  refseq_lut = NULL, refseq_name = "HXB2", pngs2o = T,
  tf_loss_cutoff = NULL, frequency_when_up = 10, exclude_vloops = F,
  included_sites = NULL, excluded_sites = NULL)
}
\arguments{
\item{aas_aln}{Alignment matrix.}

\item{aas_file}{Alignment file.}

\item{alignment_format}{Format of alignment file/s; must be one of these: \code{"fasta"}, \code{"clustal"}, \code{"phylip"}, \code{"msf"}, or \code{"mase"}.}

\item{tf_index}{TF index}

\item{tf_name}{TF name}

\item{timepoints_parser}{Timepoints parsing function generated by lassie::create.timepoint.parser()}

\item{refseq_lut_file}{Reference sequence LUT file}

\item{refseq_lut}{Reference sequence lookup table.  Optional, this could help to apply reference sequence numbering from a separate alignment that does not contain the reference sequence.}

\item{refseq_name}{Reference sequence name, used for numbering, e.g. HXB2.}

\item{pngs2o}{Switch to mark asparagines (N) in PNG motifs as O.}

\item{tf_loss_cutoff}{Threshold value (or vector of values) for including a site.}

\item{frequency_when_up}{Sites are sorted by when they first reach this value.}

\item{exclude_vloops}{Automagically add hypervariable loop sites to list of excluded sites.}

\item{included_sites}{List of included sites.}

\item{excluded_sites}{List of excluded sites.}
}
\value{
swarmtools object
}
\description{
Given a serially sampled protein sequence alignment, this will list sites where loss of transmitted-founder (TF) form exceeds a cutoff value within a given sample timepoint.
}
\details{
The sample timepoints should be part of the sequence names and it should be possible to extract them by splitting sequence names with a particular separator and taking one of the resulting fields as the timepoint label.  TF loss is computed among sequences that have the same timepoint label.  By default, the timepoint label is in the first dot-delimited field.  However, you can specify alternatives using lassie::create.timepoint.parser() to specify how fields are separated and which field to use, then passing the returned function to swarmtools using the 'timepoints_parser' option.

Timepoint labels can be in any units and are assumed to indicate time elapsed since infection.  
A mixture of different units is not advised.  
The values are used primarily when reordering sites and plotting frequency dynamics:
lassie::report.variant.frequencies()
After parsing the correct field, any leading character (e.g. "d123" or "56dpi") will be stripped and the remainder converted to a numeric value (123 and 56).

You should provide at least the name of an alignment file or an alignment matrix (via seqinr::as.alignment.matrix()) 
already in memory, and specify a percentage value (0-100) for tf_loss_cutoff.  Without these, no sites will be returned.
This is not an error state, as the returned swarmtools object could be passed to lassie::set.alignment.file() and 
lassie::set.tf.loss.cutoff().
}
\examples{
\dontrun{
  A <- lassie::swarmtools(aas_file=system.file("extdata", "CH505-gp160.fasta", 
            package="lassie"), tf_loss_cutoff=80)
  B <- lassie::swarmset(A)
}
}
