% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate.swarmset.R
\name{tabulate.swarmset}
\alias{tabulate.swarmset}
\title{List concatamers of selected sites for all working_swarm sequences in swarmset.}
\usage{
tabulate.swarmset(x)
}
\arguments{
\item{x}{swarmset}

\item{...}{not currently used by could and should be soon}
}
\description{
List concatamers of selected sites for all working_swarm sequences in swarmset.
}
\seealso{
\code{\link{swarmset}}

Other swarmset.methods: \code{\link{export.swarmset}},
  \code{\link{make.timepoint.logos}},
  \code{\link{plot.swarmset}},
  \code{\link{print.swarmset}},
  \code{\link{resample.swarmset}},
  \code{\link{summary.swarmset}}, \code{\link{swarmset}}
}

