% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vgrepl.R
\name{vgrepl}
\alias{vgrepl}
\title{Compare a vector of patterns with an input string}
\usage{
vgrepl(patterns, x)
}
\arguments{
\item{patterns}{Character vector of regular expressions to evaluate.}

\item{x}{String to search for expressions.}
}
\description{
Returns Boolean vector of length equal to the input vector that indicates which patterns match the string.
}

