% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa.colors.R
\docType{data}
\name{aa.col}
\alias{aa.col}
\title{Color Look-up-Table}
\format{An object of class \code{data.frame} with 36 rows and 4 columns.}
\usage{
aa.col
}
\description{
Used internally to depict amino acid mutations.  
To customize the color scheme, edit 'color-lut.csv', 
then uncomment the 'overwrite' argument, rebuild, reinstall, and restart.
}
\keyword{datasets}

