% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.features.R
\docType{data}
\name{env.features}
\alias{env.features}
\title{Env Feature Annotation}
\format{An object of class \code{matrix} with 856 rows and 21 columns.}
\source{
\url{http://hiv.lanl.gov/content/sequence/HIV/MAP/Env_features.xls}
}
\usage{
env.features
}
\description{
Identifies sites in HIV-1 Env associated with neutralizating antibody responses, 
whether by structural contact or site-directed mutagenesis with phenotypic assay.
}
\keyword{datasets}

