% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lassie-package.R
\docType{package}
\name{lassie}
\alias{lassie}
\alias{lassie-package}
\alias{package-lassie}
\title{Getting Help for Errors and Warnings}
\description{
Famous pooch Lassie was always there to help Timmy get help with whatever problem he was
facing. No matter what Timmy said, Lassie always understood what kind of help
was needed and how to ask for it.\cr\cr
\pkg{lassie} strives to give R users the same kind of support in finding help
for their R needs, by:
\itemize{
  \item Auto-searching for help on StackOverflow, Github, or Google and
  \item Auto-templating and posting well-formed questions to StackOverflow, if searches don't help.
}
}

