% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/make.timepoint.logos.R
\name{make.timepoint.logos}
\alias{make.timepoint.logos}
\title{Render a logo plot of concatamer forms of clones in working_swarm}
\usage{
make.timepoint.logos(x, sort_stacks = F, stacks_per_line = NULL,
  name_prefix = "logos", stratify = T, dotify = T, stack_width = 18,
  aspect_ratio = 3, logo_format = "pdf")
}
\arguments{
\item{x}{A swarmset object populated with working_swarm.}

\item{sort_stacks}{If true, reorder sites from left to right.}

\item{stacks_per_line}{If NULL, this is set to the number of selected sites; otherwise, it limits plot width.}

\item{name_prefix}{String used to identify the output files.}

\item{stratify}{If true, one logo plot is made per time-point sequenced.}

\item{dotify}{If true, one amino-acid state per site will be left blank to indicate frequency of the TF form.}

\item{stack_width}{Width of each stack, in pixels.}

\item{aspect_ratio}{Adjusts aspect ratio (width to height) of image.}

\item{logo_format}{Specifies the output image format (default is 'pdf' but can also be 'png', 'svg', or 'jpeg').}
}
\value{
A vector of names of the logo plot files generated, located in a directory removed at the end of the R session.  NB: You will need to copy this file (or these files) during run time or else lose them when the R session ends.
}
\description{
Render a logo plot of concatamer forms of clones in working_swarm
}
\examples{
\dontrun{
A <- lassie::swarmtools(aas_file=system.file("extdata", "CH505-gp160.fasta",
            package="lassie"), tf_loss_cutoff=80)
B <- lassie::swarmset(A)
logo.files <- make.timepoint.logos(B)
system(paste(paste("open", logo.files), collapse=';')
}
}
\seealso{
Other swarmset.methods: \code{\link{plot.swarmset}};
  \code{\link{print.swarmset}};
  \code{\link{resample.swarmset}};
  \code{\link{select.sequences}};
  \code{\link{summary.swarmset}}; \code{\link{swarmset}}
}

