% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.swarmset.R
\name{plot.swarmset}
\alias{plot.swarmset}
\title{Render a logo plot of concatamer forms of clones in working_swarm}
\usage{
\method{plot}{swarmset}(x, sort_stacks = F, stacks_per_line = NULL,
  dotify = F, color_scheme = "charge", ...)
}
\arguments{
\item{x}{A swarmset object populated with working_swarm.}

\item{sort_stacks}{If true, reorder sites from left to right.}

\item{stacks_per_line}{If NULL, this is set to the number of selected sites; otherwise, it limits plot width.}

\item{dotify}{If true, one amino-acid state per site will be left blank to indicate frequency of the TF form.}

\item{color_scheme}{Defines how to color the amino acids in each site.  Default is 'charge' but could also be 'monochrome', 'classic', 'hydrophobicity', or 'chemistry').}

\item{...}{can include "aspect_ratio" to adjust Aspect ratio (width to height) of image and "format" to specify the output image format (default is 'png_print' (600dpi resolution) but can also be 'png' (96dpi), 'pdf', 'svg', or 'jpeg').}
}
\value{
An explicit path to the file generated, located in a directory removed at the end of the R session.  NB: You will need to copy this file during run time or else lose it when the R session ends.
}
\description{
Render a logo plot of concatamer forms of clones in working_swarm
}
\seealso{
Other swarmset methods: \code{\link{export.fasta.swarmset}},
  \code{\link{export.swarmset}},
  \code{\link{make.timepoint.logos}},
  \code{\link{print.swarmset}},
  \code{\link{resample.swarmset}},
  \code{\link{summary.swarmset}}, \code{\link{swarmset}},
  \code{\link{tabulate.swarmset}}
}
