% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.swarmtools.R
\name{plot.swarmtools}
\alias{plot.swarmtools}
\title{Plot a cumulative distribution function of number of sites selected vs. TF loss cutoff.}
\usage{
\method{plot}{swarmtools}(x, ...)
}
\arguments{
\item{x}{Swarmtools object}

\item{...}{plot options, parsed and passed to plot(): cex, xlab, ylab, pch, lwd, main, tcl, oma, mar, mgp, xlim, ylim.  Defaults are set but you can over-write them to customize plot appearance.}
}
\description{
Plot a cumulative distribution function of number of sites selected vs. TF loss cutoff.
}
