% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.swarmtools.R
\name{print.swarmtools}
\alias{print.swarmtools}
\title{Print a table of the selected sites.}
\usage{
\method{print}{swarmtools}(x, ...)
}
\arguments{
\item{x}{Swarmtools object}

\item{...}{not currently in use but should be soon}
}
\description{
Print a table of the selected sites.
}
