% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.alignment.file.R
\name{read.alignment.file}
\alias{read.alignment.file}
\title{Read a fasta-formatted alignment file}
\usage{
read.alignment.file(aas_file = NULL, alignment_format = "fasta")
}
\arguments{
\item{aas_file}{File containing aligned protein sequences}

\item{alignment_format}{Format of alignment file/s; must be one of these: \code{fasta}, \code{clustal}, \code{phylip}, \code{msf}, or \code{mase}.}
}
\value{
A matrix that contains the alignment.
}
\description{
Read a fasta-formatted alignment file
}
\examples{
\dontrun{
aln <- read.alignment.file(system.file("extdata", "CH505-gp160.fasta", 
                                        package="lassie"))
}
}
