% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.swarmset.R
\name{resample.swarmset}
\alias{resample.swarmset}
\title{Randomly sample as many sequences as are in the given working swarmset, using the same inclusion/exclusion constraints, for a fair comparison.}
\usage{
resample.swarmset(SST, n_replicates = 100)
}
\arguments{
\item{SST}{swarmset object}

\item{n_replicates}{number of replicates}
}
\description{
Randomly sample as many sequences as are in the given working swarmset, using the same inclusion/exclusion constraints, for a fair comparison.
}
\seealso{
\code{\link{swarmset}}

Other swarmset.methods: \code{\link{export.swarmset}},
  \code{\link{make.timepoint.logos}},
  \code{\link{plot.swarmset}},
  \code{\link{print.swarmset}},
  \code{\link{summary.swarmset}}, \code{\link{swarmset}},
  \code{\link{tabulate.swarmset}}
}

