% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.swarmset.R
\name{summary.swarmset}
\alias{summary.swarmset}
\title{List names of sequences in working_swarm of swarmset.}
\usage{
\method{summary}{swarmset}(object, ...)
}
\arguments{
\item{object}{swarmset object}

\item{...}{not currently in use but should be soon
#' @seealso \code{\link{swarmset}}}
}
\description{
List names of sequences in working_swarm of swarmset.
}
\seealso{
Other swarmset.methods: \code{\link{export.swarmset}},
  \code{\link{make.timepoint.logos}},
  \code{\link{plot.swarmset}},
  \code{\link{print.swarmset}},
  \code{\link{resample.swarmset}}, \code{\link{swarmset}},
  \code{\link{tabulate.swarmset}}
}

