% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.swarmset.R
\name{summary.swarmset}
\alias{summary.swarmset}
\title{List names of sequences in working_swarm of swarmset.}
\usage{
\method{summary}{swarmset}(object, ...)
}
\arguments{
\item{object}{swarmset object}

\item{...}{not currently in use but should be soon
#'}
}
\description{
List names of sequences in working_swarm of swarmset.
}
\seealso{
\code{\link{swarmset}}

Other swarmset.methods: \code{\link{make.timepoint.logos}};
  \code{\link{plot.swarmset}};
  \code{\link{print.swarmset}};
  \code{\link{resample.swarmset}};
  \code{\link{select.sequences}}; \code{\link{swarmset}}
}

