% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/summary.swarmtools.R
\name{summary.swarmtools}
\alias{summary.swarmtools}
\title{Summarize selected sites.}
\usage{
\method{summary}{swarmtools}(S)
}
\arguments{
\item{Swarmtools}{object}
}
\value{
A data frame of selected sites, with one row per ?
}
\description{
Summarize selected sites.
}

