% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarmset.R
\name{swarmset}
\alias{swarmset}
\title{Select sequences to represent mutations in selected sites.}
\usage{
swarmset(ST, maxtol_deletion = NULL, maxtol_insertion = NULL,
  included_clone_names = NULL, excluded_clone_names = NULL,
  min_counts = 2)
}
\arguments{
\item{ST}{swarmtools object, with list of selected sites.}

\item{maxtol_deletion}{Maximum tolerated deletion length.}

\item{maxtol_insertion}{Maximum tolerated insertion.}

\item{included_clone_names}{List of included clone names.}

\item{excluded_clone_names}{List of excluded clone names.}

\item{min_counts}{Variants that occur less often than this number in the 
entire alignment are not included in swarms.  The default value (2) catches 
all recurrent mutations.}
}
\value{
A swarmset object with selected sequences in a slot called 
\code{working_swarm} with values \code{is_included==T}, 
i.e. \code{names(which(B$working_swarm$is_included))}.
}
\description{
Analyze longitudinal protein sequences for variants among selected sites.
}
\examples{
\dontrun{
  A <- lassie::swarmtools(aas_file=system.file("extdata", "CH505-gp160.fasta", 
            package="lassie"), tf_loss_cutoff=80)
  B <- lassie::swarmset(A)
}

}
\seealso{
\code{\link{swarmtools}}

Other swarmset.methods: \code{\link{export.fasta.swarmset}},
  \code{\link{export.swarmset}},
  \code{\link{make.timepoint.logos}},
  \code{\link{plot.swarmset}},
  \code{\link{print.swarmset}},
  \code{\link{resample.swarmset}},
  \code{\link{summary.swarmset}},
  \code{\link{tabulate.swarmset}}
}

