% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swarmtools.R
\name{swarmtools}
\alias{swarmtools}
\title{Identify sites under putative immune selection.}
\usage{
swarmtools(aas_aln = NULL, aas_file = NULL, alignment_format = "fasta",
  tf_index = 1, tf_name = NULL, timepoints_parser = NULL,
  refseq_lut_file = NULL, refseq_lut = NULL, refseq_name = "HXB2",
  pngs2o = F, tf_loss_cutoff = NULL, included_sites = NULL,
  excluded_sites = NULL)
}
\arguments{
\item{aas_aln}{Alignment matrix.}

\item{aas_file}{Alignment file.}

\item{alignment_format}{Format of alignment file/s; must be one of these: \code{"fasta"}, \code{"clustal"}, \code{"phylip"}, \code{"msf"}, or \code{"mase"}.}

\item{tf_index}{TF index}

\item{tf_name}{TF name}

\item{timepoints_parser}{Timepoints parsing function}

\item{refseq_lut_file}{Reference sequence LUT file}

\item{refseq_lut}{Reference sequence lookup table}

\item{refseq_name}{Reference sequence name}

\item{pngs2o}{Switch to mark asparagines (N) in PNG motifs as O.}

\item{tf_loss_cutoff}{Threshold value (or vector of values) for including a site}

\item{included_sites}{List of included sites}

\item{excluded_sites}{List of excluded sites}
}
\value{
swarmtools object
}
\description{
Given a serially sampled protein sequence alignment, this will list sites where loss of transmitted-founder (TF) form exceeds a cutoff value within a given sample timepoint.
}
\details{
The sample timepoints should be part of the sequence names and it should be possible to extract them by splitting sequence names with a particular separator and taking one of the resulting fields as the timepoint label.  TF loss is computed among sequences that have the same timepoint label.  By default, the timepoint label is in the first dot-delimited field.  However, you can specify alternatives using lassie::create.timepoint.parser().

You should provide at least the name of an alignment file or an alignment matrix (via seqinr::as.alignment.matrix()) 
already in memory, and specify a percentage value (0-100) for tf_loss_cutoff.  Without these, no sites will be returned.
This is not an error state, as the returned swarmtools object could be passed to lassie::set.alignment.file() and 
lassie::set.tf.loss.cutoff().
}
\examples{
\dontrun{
  A <- lassie::swarmtools(aas_file=system.file("extdata", "CH505-gp160.fasta", 
            package="lassie"), tf_loss_cutoff=80)
  B <- lassie::swarmset(A)
}
}

