% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/swarmtools.R
\name{swarmtools}
\alias{swarmtools}
\title{Analyze sites a protein sequence alignment for loss of the reference form.}
\usage{
swarmtools(aas_aln = NULL, aas_file = NULL, alignment_format = "fasta",
  tf_index = NULL, tf_name = NULL, timepoints_parser = NULL,
  refseq_lut_file = NULL, refseq_lut = NULL, refseq_name = "HXB2",
  tf_loss_cutoff = NULL, included_sites = NULL, excluded_sites = NULL)
}
\arguments{
\item{aas_aln}{Alignment matrix.}

\item{aas_file}{Alignment file.}

\item{alignment_format}{Format of alignment file/s; must be one of these: \code{"fasta"}, \code{"clustal"}, \code{"phylip"}, \code{"msf"}, or \code{"mase"}.}

\item{tf_index}{TF index}

\item{tf_name}{TF name}

\item{timepoints_parser}{Timepoints parsing function}

\item{refseq_lut_file}{Reference sequence LUT file}

\item{refseq_lut}{Reference sequence lookup table}

\item{refseq_name}{Reference sequence name}

\item{tf_loss_cutoff}{Threshold value (or vector of values) for including a site}

\item{included_sites}{List of included sites}

\item{excluded_sites}{List of excluded sites}
}
\value{
swarmtools object
}
\description{
Analyze sites a protein sequence alignment for loss of the reference form.
}
\examples{
\dontrun{
  A <- lassie::swarmtools(aas_file=system.file("extdata", "CH505-gp160.fasta",
            package="lassie"), tf_loss_cutoff=80)
  B <- lassie::swarmset(A)
}
}

